# Optimize humanize inflector